move join form to component
